mm, thp: consolidate THP gfp handling into alloc_hugepage_direct_gfpmask
THP allocation mode is quite complex and it depends on the defrag mode. 
This complexity is hidden in alloc_hugepage_direct_gfpmask from a large
part currently.  The NUMA special casing (namely __GFP_THISNODE) is
however independent and placed in alloc_pages_vma currently.  This both
adds an unnecessary branch to all vma based page allocation requests and
it makes the code more complex unnecessarily as well.  Not to mention that
e.g.  shmem THP used to do the node reclaiming unconditionally regardless
of the defrag mode until recently.  This was not only unexpected behavior
but it was also hardly a good default behavior and I strongly suspect it
was just a side effect of the code sharing more than a deliberate decision
which suggests that such a layering is wrong.

Get rid of the thp special casing from alloc_pages_vma and move the logic
to alloc_hugepage_direct_gfpmask.  __GFP_THISNODE is applied to the
resulting gfp mask only when the direct reclaim is not requested and when
there is no explicit numa binding to preserve the current logic.

This allows for removing alloc_hugepage_vma as well.

Link: http://lkml.kernel.org/r/20180925120326.24392-3-mhocko@kernel.org
Signed-off-by: Michal Hocko <mhocko@suse.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: "Kirill A. Shutemov" <kirill@shutemov.name>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Stefan Priebe - Profihost AG <s.priebe@profihost.ag>
Cc: Zi Yan <zi.yan@cs.rutgers.edu>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>