debugobjects: scale the static pool size
The current value of the early boot static pool size is not big enough for
systems with a large number of CPUs with timer or/and workqueue objects
selected.  As a result, systems having 60+ CPUs with both timer and
workqueue objects enabled could trigger "ODEBUG: Out of memory.  ODEBUG
disabled".  Fix this by computing it according to CONFIG_NR_CPUS and
CONFIG_DEBUG_OBJECTS_* options.

Link: http://lkml.kernel.org/r/20181121021157.3061-1-cai@gmx.us
Signed-off-by: Qian Cai <cai@gmx.us>
Acked-by: Waiman Long <longman@redhat.com>
Cc: Yang Shi <yang.shi@linux.alibaba.com>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>